backend/pyproject.toml
[project]
name = "narrative-game-base"
version = "0.2.0"
description = "A multi-agent narrative game system using LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.0.10",
    "langchain-groq>=0.1.0",
    "pydantic>=2.0.0",
    "langchain-core",
    "python-dotenv>=1.2.1",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]

[tool.uv]
python-version = "3.11"
